+++ packages/adminui-vue/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `UaclRlsInput` component. Recommended to use for managing access to some entity with aclRls mixin
   [See documentation here](https://git-pub.intecracy.com/unitybase/ubjs/-/blob/master/packages/adminui-vue/components/UAclRlsInput.vue)
+- `addCollectionItemWithoutDefaultValues` - new Vuex action - such as `addCollectionItem`
+but does not fetch default params except ID
 
 ### Changed